Move Const to Free
