+++ modules/core/arrow-mtl/src/main/kotlin/arrow/mtl/instances/const.kt
@@ -2,10 +2,10 @@ package arrow.mtl.instances
 
 import arrow.Kind
 import arrow.core.Option
-import arrow.data.Const
-import arrow.data.ConstOf
-import arrow.data.ConstPartialOf
-import arrow.data.fix
+import arrow.free.Const
+import arrow.free.ConstOf
+import arrow.free.ConstPartialOf
+import arrow.free.fix
 import arrow.free.instances.ConstTraverseInstance
 import arrow.instance
 import arrow.mtl.typeclasses.TraverseFilter
